eclass: use skylark for toolchain configuration
In preparation for upgrading Bazel, we need to define the toolchain in
Skylark and not in a config proto.

This version of the eclass is compatible with Bazel 0.24.1 (current)
and Bazel 3.2.0 (future). This is one less moving part for the main
TFLite uprev.

Note: This addresses the problems from the revert in
I50e7dc02b9c832e221d9ca7163b4b7d9c918ac4f. Previosly it was assumed all
the features from the original eclass were unused since they weren't
enabled. In fact, they get enabled from the main bazel.eclass which
hardcores the 'opt' build configuration. The original change is
patchset #1, and the fixes are in the following patchsets.

BUG=chromium:1102795,b:159987197
TEST=emerge-betty-pi-arc chromeos-base/drivefs sci-libs/tensorflow
     USE="-cros-debug" emerge-betty-pi-arc chromeos-base/drivefs \
                                           sci-libs/tensorflow

Change-Id: I69abf4ba3d55d66b547d58766f79baa6bbc6af4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2292412
Tested-by: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Sergei Datsenko <dats@chromium.org>
Commit-Queue: Jim Pollock <jmpollock@chromium.org>
Auto-Submit: Jim Pollock <jmpollock@chromium.org>