afdo: Enable AFDO profile for kernel 4.14
We have been using kernel AFDO profile on auron_yuna for a couple
of months now. There's not enough data to proof the performance
gain with AFDO because limited number of boards available. But
since it's safe and no proof of slowing down, we should go ahead
and enable AFDO for all 4.14 kernel.

BUG=b:154543384
TEST=Will use tryjobs and CQ

Change-Id: I2709fb7eff9c5c45cdde837fb3fd04857a35a578
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2295657
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Tiancong Wang <tcwang@google.com>
Commit-Queue: Tiancong Wang <tcwang@google.com>