dedede: Add madoo coreboot configuration
Create a new coreboot configuration for the madoo variant
of the waddledoo reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.1.1).

BUG=b:161191394
TEST=FW_NAME=madoo emerge-dedede coreboot chromeos-bootimage
Ensure that image-madoo.*.bin are created

Change-Id: I1dee76d51a6ffe67dcaaeb4afa0e059fe2019e8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2297359
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Henry Sun <henrysun@google.com>
Commit-Queue: Henry Sun <henrysun@google.com>
Tested-by: Henry Sun <henrysun@google.com>