chromeos-fonts: isolate temporary bind mounts
It's dangerous to add public mounts to well-known paths in a ${SYSROOT}
and then expect to be able to unmount them cleanly later. As ${BUG}
shows, we might well see EBUSY if others are starting to use it.

Use a mount namespace with a private bind mount instead.

Doing this all with tools linux util-linux 'unshare' got fairly
complicated with all the shell bootstrapping around it. It turns out we
can simplify this a lot by just leveraging platform2_test.py for most of
the QEMU handling. The only thing missing is custom bind mounts, so we
wrap in minijail0 too.

Note that MS_BIND and MS_PRIVATE aren't currently available for easy
parsing in minijail0, due to quirks in its "gen_constants" scripts.

BUG=chromium:1049710
TEST=build repeatedly (for SDK, and for ARCH={x86,arm} targets) and
     ensure that root mount namespace remains unchanged throughout

Change-Id: I667ed87cc8e0c5d46bf8504bf7edfd1c44aa1363
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2298218
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>