dedede: Add magolor coreboot configuration
Create a new coreboot configuration for the magolor variant
of the waddledee reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.1.1).

BUG=b:158540772
TEST=FW_NAME=magolor emerge-dedede coreboot chromeos-bootimage
Ensure that image-magolor.*.bin are created

Change-Id: I59f7086c1ba1559f3493a9ad082732d48b1f2667
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2301692
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Henry Sun <henrysun@google.com>
Commit-Queue: Henry Sun <henrysun@google.com>
Tested-by: Henry Sun <henrysun@google.com>