iHD: Re-enable the CMRT library.
This CL allows building the iHD driver with the CMRT library as long as
the ihd_cmrtlib USE flag is set. This was originally done in CL:2143233.
It was disabled by CL:2303890 with the intention of re-enabling it in
its own CL for bisectability.

BUG=b:161469007
TEST=./build_packages --board=dedede with USE=ihd_cmrtlib generates
     /usr/lib64/libigfxcmrt.so

Cq-Depend: chromium:2303899
Change-Id: Iec552bfaa47993c857f8ee74239e736bfc80073b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2303128
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>