chromium-source: pull out depot_tools variables into the eclass
These constants should be set to these fixed values all the time,
so pull them into the eclass to share them among users.  We drop
the ability to customize these as no one ever has, and we don't
really want them to in the first place.

This also fixes the eclass to stop looking for depot_tools in the
~/ homedir which is the old deprecated path.

BUG=None
TEST=CQ passes

Change-Id: I91935cd606f1a86dceff95573caa4850dc45c9cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2336285
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>