volteer: Add eldrid coreboot configuration
Create a new coreboot configuration for the eldrid variant
of the volteer reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.1.1).

BUG=b:162115131
TEST=FW_NAME=eldrid emerge-volteer coreboot chromeos-bootimage
Ensure that image-eldrid.*.bin are created

Cq-Depend: chromium:2323112
Change-Id: I6b1ebe58f62f5d82aab502fa1fbf138b7b3a6bbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2319946
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: caveh jalali <caveh@chromium.org>
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Tested-by: YH Lin <yueherngl@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>