volteer: Add todor coreboot configuration
Create a new coreboot configuration for the todor variant
of the volteer reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.1.1).

BUG=b:162110806
TEST=FW_NAME=todor emerge-volteer coreboot chromeos-bootimage
Ensure that image-todor.*.bin are created

Change-Id: I8da606003d43e095164b9b3ed3ba0832cab4a086
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2318283
Tested-by: YH Lin <yueherngl@chromium.org>
Reviewed-by: Zhuohao Lee <zhuohao@chromium.org>
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Commit-Queue: YH Lin <yueherngl@chromium.org>