chromeos-zephyr: add zephyr-based ec firmware target
This ebuild builds and installs zephyr EC firmware into
/firmware/${target}/ec.bin, for all zephyr firmwares defined in
cros_config.

The guts of the logic is in cros-zephyr.eclass: this is so that we can
define multiple zephyr ebuilds on a per-version basis.  We'll have to
add a USE flags selection logic to cros-firmware.eclass for each
program in a follow-up CL.

BUG=b:157245202
TEST=emerge-betty chromeos-zephyr-2_3, observe
     /build/betty/firmware/scarlet/ec.bin

Change-Id: Id18d7584d5ddf9603d3d0ff1181f0cf101f3fe97
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2300335
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>