sys-boot/grub: fix warning caused by the gptpriority patch
The patch was causing this warning:

    /var/tmp/portage/sys-boot/grub-2.02-r2/work/grub-2.02/grub-core/commands/gptpriority.c:67:18: warning: format '%lu' expects argument of type 'long unsigned int', but argument 4 has type 'long long unsigned int' [-Wformat=]
                      (part->gpt_attrib >> 48) & 0xfULL);

Presumably the patch was originally compiled for x86, where uint64 is
an `unsigned long long int`, but on amd64 it's an `unsigned long int`.

I don't think that specifying the exact type of the 0xf constant is
needed here, so I just removed the `ULL` suffix so that it should work
everywhere.

This warning was hidden in the noise because grub has a lot of
warnings, so the ebuild sets `--disable-werror`.

BUG=None
TEST=sudo emerge grub, verify that the warning is gone

Change-Id: Idb3dded8c27f86b8c8ead028559394af07ec6c81
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2332894
Tested-by: Nicholas Bishop <nbishop@neverware.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>