chromeos-bootimage: Add tianocore if possible
Some boards have only 1MB of RW_LEGACY space. It is a pain to figure out
whether we actually do have space apriori. We can't put a USE flag in the
board since the space available to specific models can vary (e.g. hatch).
We could use cros_config to control this, but it seems like a lot of
effort.

It is easier to just add it if we can. It is of course unfortunate in
that the tianocore USE flag effectively ends up being ignored.

BUG=chromium:1114872
TEST=USE=tianocore FW_NAME=akemi emerge-hatch chromeos-bootimage
See that it now builds correctly

Change-Id: Ic469a002cd68e2abc29ec734b14a43fa1b4f40f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2352019
Tested-by: Simon Glass <sjg@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>