chromeos-bootimage: show size info for assets and available space
With the introduction of Groot, one challenge has been fitting
assets (pre-generated images of i18n strings and other sprites)
in flash images across all boards built in ToT.  This CL provides
some feedback of available space as part of the chromeos-bootimage
build process.  The printed messages are purely informational,
due to the limitation that it is impossible to ascertain the final
image size without actually performing the `cbfstool add` steps.

Move the `add_assets` step to the end of image preparation.
At this point, only the payload and assets remain to be added
to the image.

Calculate remaining space in the image, and estimate the size
these objects will take on flash, taking LZMA compression into
account.  If the size exceeds available space, generate a warning
of how many KB over the limit is projected to be used.

Additionally, show the size available for assets, and the
projected size to be used by assets.

Example output:
  * assets (RO): 4355 KB (2994 KB free) puff
  * WARNING: RO estimated 1590 KB over limit puff
  * assets (RW): 4355 KB (3328 KB free) puff
  * WARNING: RW estimated 1257 KB over limit puff

BUG=b:146399181, b:160362285
TEST=Build locally with various bmpblk settings;
     check that ebuild informational output is accurate

Change-Id: I0010d1bfaa5aadd33f25a9af49331f4ea08115ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2353733
Tested-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Commit-Queue: Joel Kitching <kitching@chromium.org>