cros-workon: fix partial checkout fallbacks
This logic was written when we only supported a single project, so
falling back from a single failure was easy.  Now that we support
checking out multiple repos but only one fails, the fallback logic
tries to check them all out again which fails as git doesn't let
you (re)clone a repo on top of an existing one.

If we see a .git dir, assume it's been checked out above.  If git
had failed earlier, it should have removed the .git.

BUG=None
TEST=emerging chromeos-base/tast-local-tests-crosint with only one repo out of sync works again

Change-Id: Ie86a7e4a837ede4f2a89c9694b2715b0bbf3bf2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2357251
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>