chromeos-base: Add rule for Semtech SX9310 and SX9311
Add a rule to create a link to the proximity sensor and enable the
channel in use. The name of the link reflects which antenna the sensor
is protecting. This rule only runs when the sysfs attribute 'label' is
present, which so far is only supported on DT based devices.

BUG=b:162279829
TEST=See /dev/proximity-wifi exists on Lazor

Change-Id: Ib391f743420601df80310b91f376b1193e59cf7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2359484
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>