hostapd: Enable CONFIG_NO_RANDOM_POOL.
When hostapd starts, it tries to read some bytes from /dev/random to
see if there's enough entropy. If not, it will reject the next key
handshake.

On Whirlwind test APs, the tpm-rng has slow entropy generation. When we
run WiFi tests, which spawn/kill hostapd frequently, it is possible that
the entropy is almost all drained. Then the tests which need key
generation, e.g. WPA tests, will fail due to lack of entropy.

The old image is OK as it has crrev.com/c/397721 which works around the
case.

As it is used only for testing, we don't quite care about the entropy,
disable the random_pool_ready check so that hostapd will just read from
/dev/urandom as random source instead.

BUG=chromium:1114912
TEST=Run tast.wificell-func-stable and monitor entropy_avail on router.
It should not be drained to near 0 now.

Change-Id: I6f3573c02bc49d02f342f8099330d71f35effb48
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2374426
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Commit-Queue: Yen-lin Lai <yenlinlai@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>