arc-adbd: Change the seccomp policy file name for ARCVM
This makes the file name consistent with the Upstart conf file name
(arcvm-adbd.conf). Also, arcvm-adbd.conf is referring to the policy
file as arcvm-adbd-seccomp.policy, so the current newins code here
is causing an error.

This is effectively a reland of
https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2028508
with the dest file name tweak.

BUG=b:166314117
TEST=emerge-eve-arc-r arc-adbd, build_packages for kukui-arc-r

Cq-Depend: chromium:2382307
Change-Id: Ide27cf97ac796367c964cab3c08f952c2bb865ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2381561
Tested-by: Yusuke Sato <yusukes@chromium.org>
Commit-Queue: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>