Make (test) images without Python 2.7 possible
Modifies a handful of top-level packages so that they only depend
on python2-exclusive packages when the `python_targets_python2_7`
USE flag is enabled. Tweaks chromeos/package.use so that this ought
to make no difference by default.

Removes python2-only dependencies from the factory and factory-deps
ebuilds entirely because (I'm told) they aren't needed any more.

The motivation here is that the stateful partition has grown too
large for certain devices, and we need to trim it down so that
we can reliably flash the Gale routers used for automated WiFi
testing. Since we're pretty sure Python 2.7 is not being used
for anything on these routers, it's an easy ~143 MiB reduction.

See also: crrev/i/3193765 for the testbed-ap changes that this
CL enables.

I have verified that with this CL alongside the changes from
crrev/i/3193765 the image no longer has /usr/local/bin/python2.7
or directory /usr/local/lib/python2.7/. I have also verified
that in the without that extra CL the interpreter and libraries
are still present, so this CL appears to have no effect on its
own, as intended.

BUG=chromium:1104412
TEST=built and inspected images; tested Gale without Python 2.7

Change-Id: I0cf630981747398473e2635192f6487c8c52ae6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2332476
Tested-by: Will Donnelly <wgd@google.com>
Commit-Queue: Will Donnelly <wgd@google.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Yilin Yang (kerker) <kerker@chromium.org>