media-libs/libcamera: Do not use CROS_WORKON_OUTOFTREE_BUILD
As per Meson bug [1]:

"When meson tries to identify the C/C++ compiler and linker, it runs
various arguments checks depending on the compiler and linked. For
LLVMDynamicLinker, it runs x86_64-pc-linux-gnu-clang
-Wl,--allow-shlib-undefined to check if the linker supports
--allow-shlib-undefined. This invokes the linker with -o a.out. There is
no source or object file specified on the command line, so a.out should
not be written to. However, the linker seems to still open a temporary
file in write mode (a.out.tmp6bdf1a0).

When 'meson setup' is run from the source directory with the build
directory as an argument, it doesn't change to the build directory to
run this step. There is thus an attempt to write in the source
directory, which portage (on Chrome OS) caught as it sandboxes the
build."

For this reason, if CROS_WORKON_OUTOFTREE_BUILD is used, the temporary
file ends up being in the source directory which is prohibited by the
sandbox.  As a workaround, use only CROS_WORKON_INCREMENTAL_BUILD and
set ${BUILD_DIR} manually, which allows incremental builds, but at the
cost of a source tree copy, which should be small in this case, because
the project is relatively small.

[1] https://github.com/mesonbuild/meson/issues/7676

BUG=b:166747570
TEST=emerge-soraka libcamera

Change-Id: Ia002d102803ceadb46ec38218cb9ff61592b3ff6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2385836
Tested-by: Tomasz Figa <tfiga@chromium.org>
Auto-Submit: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>