puff: Add dooly coreboot configuration
Create a new coreboot configuration for the dooly variant
of the puff reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.1.2).

BUG=b:155261464
TEST=FW_NAME=dooly emerge-puff coreboot chromeos-bootimage
Ensure that image-dooly.*.bin are created

Change-Id: Ibc599cc45b29872eb081b64a8c5127e7870756f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2390240
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Tested-by: Tony Huang <tony-huang@quanta.corp-partner.google.com>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Sam McNally <sammc@chromium.org>