sys-apps/fwupd: Enable fwupd daemon mode
This change makes the fwupd ebuild to compile the fwupd daemon and adds
an upstart script to start it. The daemon exposes D-Bus interface for
clients to communicate with. We also changed many of the fwupdtool
invocations to fwupdmgr calls that interact with the running daemon.

BUG=b:167681114
TEST=emerge-sarien fwupd; start fwupd

Change-Id: I5472ef6c8a513f1bb1867768bf617d10cb7a76d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2391972
Tested-by: Daniel Campello <campello@chromium.org>
Commit-Queue: Daniel Campello <campello@chromium.org>
Reviewed-by: Benson Leung <bleung@google.com>