Add kernel OWNERS
The intention here is not to provide any additional restrictions [1], but to
(a) informatively point people looking for owners (e.g., in sys-kernel/)
(b) allow kernel folks to OWN eclass/cros-kernel2.eclass and a few misc
    things in eclass/cros-kernel/ (which otherwise require general
    ebuild/portage-stable OWNERS approval)

Existing ("general ebuild") OWNERS are still valuable for many ebuild
questions, but cros-kernel2.eclass is especially filled with
kernel-domain logic.

The OWNERS list is a likely-biased (feel free to fix that!) filtering of
significant Author or Review lines in the git history.

[1] I don't use 'set noparent' anywhere.

BUG=none
TEST=none

Change-Id: Id8a02a7574e646ceac74a3c73efdda49319f3c30
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2392862
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>