cups-filters: Activate foomatic_shell
This patch replaces a shell used by cups-filters with a dedicated
mini-shell called foomatic-shell. One of the cups-filters tools,
foomatic-rip, uses this shell to execute scripts included in some PPD
files. We want to use foomatic-shell instead of standard shell to limit
set of commands that can be executed from provided PPD file. The goal is
to prevent execution of unauthorized tools. foomatic-shell was
implemented in CL:2271389.

BUG=chromium:1073063
TEST=on atlas with platform_PrinterPpds.all

Change-Id: I5d68cd10e3b90f4d260ead7de752e9e3a8e29275
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2395156
Tested-by: Piotr Pawliczek <pawliczek@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Piotr Pawliczek <pawliczek@chromium.org>