net-libs/librmnetctl: Install libs to lib64 on 64-bit builds
We need to use the right lib vs. lib64 directory depending on if we're
installing for 64-bit or 32-bit userspace. The Makefile in this package
only uses $(prefix)/lib as the default, so let's override it so we get
the libs in the right place.

BUG=b:168256890, b:168256711
TEST=emerge-trogdor64 net-libs/librmnetctl

Change-Id: Ic7cdfad029576be908ad597a07194339964738be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2404953
Tested-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Auto-Submit: Stephen Boyd <swboyd@chromium.org>