perfetto: Disable some warnings.
This temporarily disables -Wsign-conversion and -Wshorten-64-to-32
for perfetto. An upcoming version of the Clang has gotten stricter
with those warnings and perfetto is not building with it.  This patch
will allow us to continue testing the new compiler while the perfetto
issues are fixed.

BUG=chromium:1131186
TEST=Tested locally with new compiler; without this patch perfetto
fails to build (for Bob); with this patch it builds.

Change-Id: Iaf3f742a5d017099925c2722fc5b5f5426c38a91
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2425271
Tested-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Manoj Gupta (OoO) <manojgupta@chromium.org>
Commit-Queue: Caroline Tice <cmtice@chromium.org>
Auto-Submit: Caroline Tice <cmtice@chromium.org>