sys-kernel/linux-headers: move non-upstream VP9 headers to their own file
These controls are not in mainline and we will introduce the actual
upstream controls once they are stabilized. In order to prepare for
this, move the old controls to their own file, which we include from
v4l2-controls.h to make sure client behavior is not disturbed.

Also remove vp9-specific pointer unions and control type enums, which
are not upstream and not used by Chromium.

BUG=b:161415741
TEST=emerge-kukui linux-headers

Change-Id: I9ef370da352e9dc4e5584856665b94f1de0f554c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2425788
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>