volteer: Add boldar coreboot configuration
Create a new coreboot configuration for the boldar variant
of the volteer reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.3.0).

BUG=b:162202257
TEST=FW_NAME=boldar emerge-volteer coreboot chromeos-bootimage
Ensure that image-boldar.*.bin are created

Cq-Depend: chrome-internal:3285729
Change-Id: Iea7dedb403d1a153e01aecabc7be7b10e88245ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2426123
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Tested-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Commit-Queue: Tim Wawrzynczak <twawrzynczak@chromium.org>