sane-backends: install test backend on test images
Unmask the use flag for the test backend, and add logic to the
sane-backends ebuild to move the lib into /usr/local/[libdir] so that it
will only be installed on test images.

Add an ebuild that creates symlinks to that lib in sane's libdir.
Install that ebuild on test images.

BUG=b:166646255
TEST=scanimage -nA -d test

Change-Id: I694536c310b48160b57e7b2a4923f811c2c2344b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2432545
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>