volteer: Add lindar coreboot configuration
Create a new coreboot configuration for the lindar variant
of the volteer reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.3.0).

BUG=b:161089195
TEST=FW_NAME=lindar emerge-volteer coreboot chromeos-bootimage
Ensure that image-lindar.*.bin are created

Change-Id: I9bfb8a5dd521c4a841c7d6d6ecba1ee285675414
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2434177
Reviewed-by: YH Lin <yueherngl@chromium.org>
Reviewed-by: caveh jalali <caveh@chromium.org>
Tested-by: rasheed.hsueh <rasheed.hsueh@lcfc.corp-partner.google.com>
Commit-Queue: YH Lin <yueherngl@chromium.org>