cros-kernel2: Add lockstat USE flag
Add a USE flag to enable lock statistics in the kernel. This helps find
lock contention. Throw it into the lockdebug use flag too in case it
helps find out problems while debugging locks, but keep it as a
different USE flag so that the overhead of lockdep isn't required to see
some stats for locking.

BUG=None
TEST=echo 1 >/proc/sys/kernel/lock_stat && \
     less /proc/lock_stat

Change-Id: Ib5ff7f7b753f4ab6e7bd19d677e46373c1563586
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2447172
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>