volteer: Add elemi coreboot configuration
Create a new coreboot configuration for the elemi variant
of the volteer reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.3.0).

BUG=b:170604353
TEST=FW_NAME=elemi emerge-volteer coreboot chromeos-bootimage
Ensure that image-elemi.*.bin are created

Change-Id: I8f9d833466a757ba786c551bda952f18025bbf13
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2462860
Tested-by: Wisley Chen <wisley.chen@quanta.corp-partner.google.com>
Reviewed-by: YH Lin <yueherngl@chromium.org>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: YH Lin <yueherngl@chromium.org>