libchrome: optimize for speed
This is now the 8th hottest package across all CrOS, consuming 0.22% of
fleetwide cycles. In accordance with go/cr-os, this should be optimized
for speed.

It was shelved previously because I thought doing so would cost 4MB or
so. As it turns out, the majority of that was in debuginfo: the actual
rootfs size cost appears to be a bit under 2MB (<2% of the total 120MB
or so of installed size).

BUG=None
TEST=`emerge-eve libchrome`; compared `equery f`

Change-Id: Iaa8cd8b4ee9bce9ab1e8e96d608e1d40ada1e516
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2466673
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>