sys-fs/udev: Plug a leak in sd_device_set_sysattr_value()
We use this API a little bit in powerd (see Udev::SetSysattr() and how
it calls the udev library API). I was trying to write non-existent
sysfs attributes and that seemed to cause random corruption in systemd.

It looks like this API leaks so we need to pick a few patches from
upstream that plug the leak. See
https://github.com/systemd/systemd/issues/6040 for more details. I've
picked two patches from there and applied them to the ebuild and the
leak seems to be gone.

BUG=b:165322407
TEST=Use CL:2462704 on Lazor, see that it doesn't crash powerd

Change-Id: Ice2ff6a6dc2dcf83092b69e82ddfcac35ed215ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2467003
Tested-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>