ml: move tclib model file installation to tclib's ebuild
We used to put the model installation in mlservice's ebuild because we
thought the unit test can not see the model files in the standard
location. It turns it can. So it is better to install the model files
in its own ebuild.

BUG=none
TEST=emerge-nocturne libtextclassifier
TEST=FEATURES=test emerge-nocturne ml

Cq-Depend: 2473178
Change-Id: I42186767d231ad1c045854f964faee757835781b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2472978
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Commit-Queue: Honglin Yu <honglinyu@chromium.org>
Tested-by: Honglin Yu <honglinyu@chromium.org>