Reland "Bootstrap dev-lang/rust-1.47.0 using dev-lang/rust-bootstrap"
crrev.com/c/2399400 introduced the rust-bootstrap package, which
bootstraps rustc using a trusted compiler. This change modifies
the rust-1.47.0 ebuild so that it builds using the rust-bootstrap
compiler, rather than a binary compiler release from upstream.

Original Change-Id: I800985beb1c5d9b339658c5db8e96cd3fb9c39c4

Relanding because this was landed with a broken readline upgrade that
couldn't be deployed, so this had to be reverted.

BUG=chromium:1122720
TEST=sudo emerge dev-lang/rust

Change-Id: I0716307d6985ba42f682638443f31a639112f9a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2533209
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>