Reland "rust: migrate to EAPI 7"
This migrates the dev-lang/rust ebuild to use EAPI 7 instead of the
previous EAPI 5. EAPI 7 adds support for BDEPEND, which we intend
to use to get a build-time dependency on rust-bootstrap.

Original Change-Id: I5f5b802e8df3a32a909028f45e5ce7acef8f6cc6

Relanding because this was landed with a broken readline upgrade that
couldn't be deployed, so this had to be reverted.

BUG=chromium:1122720
TEST=sudo emerge --unmerge rust && sudo emerge rust

Change-Id: Ie07e3b3cec36117f1470fa2369a7a7a144f6d3c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2533208
Tested-by: George Burgess <gbiv@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>