Reland "rust: fix rustfmt crashing on tuple.0.0 notation"
This CL adds a cherrypick to fix crashes noted in the below bug. For
reasons mentioned in the bug, simply cherrypicking this isn't enough: we
also need it to actually _apply_ to rustfmt. Hence, this modifies
vendored sources (which angers `cargo`, which recommends the use of
`[replace]` or `[patch]` instead, but those alternatives separate
checkouts of the source tree we want to override. keeping a single,
targeted patch for all of this seems better than uploading another
tarball/etc.)

Original Change-Id: I0b2026db6803420f9cf4ae9863d8ca0ed94acc46

Relanding because this was landed with a broken readline upgrade that
couldn't be deployed, so this had to be reverted.

BUG=chromium:1141457
TEST=`cargo fmt` on platform2/sirenia

Change-Id: Ie86108b38f32e534e70344267595d4acbfa00607
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2533210
Tested-by: George Burgess <gbiv@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>