volteer: Add volteer2_ti50 coreboot configuration
Create a new coreboot configuration based on the volteer2 variant.

For the Ti50 project, a number of volteer2 devices have been reworked to
replace the H1 TPM chip with a Ti50.  This is done to be able to develop
and test the Ti50 firmware, before we have the first actual Chrome OS
device to use this TPM chip.

A few changes to Volteer coreboot is required for the rework, mainly
because the Ti50 is connected on an I2C bus, whereas the old H1 TPM was
connected via SPI.

This CL is based on crrev.com/c/2255286

BUG=b:169526865
TEST=gen_config, tested by YH Lin together with CL:*3368294

Change-Id: I3421da99a5a616a4eec8473f07ae30480ed293c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2506058
Tested-by: Jes Klinke <jbk@chromium.org>
Commit-Queue: YH Lin <yueherngl@chromium.org>
Reviewed-by: YH Lin <yueherngl@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@google.com>