volteer: Add voema coreboot configuration
Create a new coreboot configuration for the voema variant
of the volteer reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.3.0).

BUG=b:171755775
TEST=FW_NAME=voema emerge-volteer coreboot chromeos-bootimage
Ensure that image-voema.*.bin are created

Change-Id: I1e120a0f2a48f81792457368155700a6fc37f6a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2508445
Reviewed-by: David Wu <david_wu@quanta.corp-partner.google.com>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Tested-by: David Wu <david_wu@quanta.corp-partner.google.com>
Commit-Queue: Tim Wawrzynczak <twawrzynczak@chromium.org>