rmtfs: Depend on syslog and explain start conditions
If we happen to try to start rmtfs too early in the boot process we
can get a strange message in the log that looks like:
  init: rmtfs main process (...) terminated with status 251

It's _really_ hard to get any info out about what happened, but after
a bunch of augmentation of minijail0 I found that it was trying to
bind mount "/dev/log".  That isn't there until syslog starts.  Let's
add a precondition for that.

While we're at it, let's explain our start conditions including
reminding that verify_fsg blocks us from starting.

BUG=b:162465221
TEST=w/ all patches, WiFi starts a bit earlier

Change-Id: Ia6610e4e0c1159bc252031a7089dfab03882ea45
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2544915
Tested-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>