mesa: sync EAPI and patch application between the ebuilds
The EAPI in the 9999 ebuild was updated to 7, but that change wasn't
made in the stable ebuild (and the ebuild is blacklisted so it didn't
happen automatically.)

To actually use EAPI=7, uses of the deprecated epatch have been
replaced with eapply.

Additionally, both the 9999 and stable ebuilds had an "apply patches"
section for applying multiple patches, but as far as I can tell this
is unused since SRC_PATCHES is not set. So, I just removed it rather
than changing to use eapply. (I also checked the gentoo ebuild to make
sure this section doesn't exist to keep the ebuilds similar, but it's
not present in the current gentoo ebuilds.)

BUG=none
TEST=Build with `USE="$USE -video_cards_radeon -video_cards_nouveau" emerge-amd64-generic mesa` (disabling radeon and nouveau was needed to avoid an unrelated error from a libdrm version check)

Change-Id: I5918158ccee9b8bc7334d82cd03fc9de15184997
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2513724
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Ilja H. Friedel <ihf@chromium.org>
Commit-Queue: Ilja H. Friedel <ihf@chromium.org>