cros-kernel2: Set OBJCOPY and STRIP when building with clang
We need to set OBJCOPY to be llvm if we're building the arm64 vdso32
with ld.lld. Otherwise we get a build error like this:

 OBJCOPY arch/arm64/kernel/vdso32/vdso.so
 aarch64-cros-linux-gnu-objcopy: arch/arm64/kernel/vdso32/vdso.so: Not enough room for program headers, try linking with -N
 aarch64-cros-linux-gnu-objcopy:arch/arm64/kernel/vdso32/vdso.so[.hash]: Bad value

This is because OBJCOPY=aarch64-cros-linux-gnu-objcopy and it is a
binutils program that we're mixing with ld.lld. Historically this wasn't
a problem; we've been building the 32-bit vdso with clang and linking it
with bfd per the kernel's Makefile but once we get commit 0fa97e9403c7
("arm64: vdso32: Allow ld.lld to properly link the VDSO") we start using
ld.lld to link and objcopy stops working.

Let's set OBJCOPY and STRIP to be llvm tools when we're compilng the
kernel with clang. This should work just as well and is The Future™ so
we might as well live in it. Note, we really should be using LLVM=1 on
the kernel make commandline here but that feature isn't backported to
v4.4, only v4.19 stable trees thanks to Nick! Once we only support
v4.19+ kernels we can start passing LLVM=1 on the commandline and avoid
having to do this sort of stuff, or we can backport the LLVM=1 feature
to v4.4 kernels and start using LLVM=1 all the time.

Primes updated with:

  sed -i '/coolest prime/s:151:157:' sys-kernel/chromeos*/*9999.ebuild

BUG=None
TEST=Build v5.10-rc2 kernel for Trogdor

Change-Id: Iffdc55064bef8fc1bf11817dfa3371a1015214c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2515084
Tested-by: Stephen Boyd <swboyd@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>