cups-filters: Fix foomatic-rip infinite loop
If the foomatic rip command is not present in the PPD file, the foomatic
rip filter runs cat by default to forward stdin to stdout. However,
if input is provided through a file instead of stdin, then this file
needs to be dup'd to stdin for the command to work properly.

Also, the next file needs to be selected each time an input file is
processed by calling strok_r in order to prevent an infinite loop.

BUG=none
TEST=foomatic-rip --ppd=ppd.ppd file.pdf

Change-Id: I86761437c3b7ff6fd37f88d74732bf600adb3913
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2519309
Tested-by: Pranav Batra <batrapranav@chromium.org>
Reviewed-by: Piotr Pawliczek <pawliczek@chromium.org>
Auto-Submit: Pranav Batra <batrapranav@chromium.org>