target-chromium-os: Add fingerprint_study cond depend
This was previously done in chromeos-bsp-nocturne-private,
but was forgotten in subsequent fingerprint boards.
This tool does not need biod (the USE flag or daemon) to run
properly.

We choose to place it in the base over dev or test because we
want to create minimal images for the chromebooks that
conduct the study.
Furthermore, if the package was pulled in via dev or test, the
upstart script would not run automatically, since those package
files are contained within /usr/local.

Note, when fpstudy is enabled, python3 is pulled into the
base image.

BUG=b:144061947
TEST=cros-workon-hatch start virtual/target-chromium-os
     USE=fpstudy ./build_packages --board=hatch
     ./build_image --board=hatch base

Change-Id: Id52fa7d7e97aaed9ce18c8e1f64ca6a71bc574ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2522099
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Craig Hesling <hesling@chromium.org>
Tested-by: Craig Hesling <hesling@chromium.org>