sys-boot/chromeos-bootimage: Fix size estimation for payload ELF
The check_assets() code needs to estimate the size of the compressed
payload in CBFS, because it runs before the payload has been added.
However, the approach taken to just XZ-compress the full ELF file is
wrong: it accounts for the compression but not for the fact that ELF
files tend to contain a lot of extra debug information and cbfstool only
copies in raw program segment bytes. This leads to a large
overestimation of the size taken up by the payload, often resulting in
warnings when there is more than enough space left in the actual image.

This patch fixes the problem by extracting raw program bytes with
objcopy for the purposes of this estimation.

BUG=None
TEST=emerge-trogdor chromeos-bootimage doesn't show warnings anymore.

Change-Id: I1241e1c49c249bbdd0929f5b9ca2acb594388904
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2522107
Reviewed-by: Bob Moragues <moragues@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>
Commit-Queue: Julius Werner <jwerner@chromium.org>
Tested-by: Julius Werner <jwerner@chromium.org>